# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CTEM Platform API — Continuous Threat Exposure Management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tenants:
    get:
      operationId: listTenants
      tags: [tenants]
      summary: List all tenants (super admin only)
      responses:
        "200":
          description: List of tenants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tenant"
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Create a new tenant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Tenant created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
  /tenants/{tenantId}:
    get:
      operationId: getTenant
      tags: [tenants]
      summary: Get tenant by ID
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tenant details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
    patch:
      operationId: updateTenant
      tags: [tenants]
      summary: Update tenant
      parameters:
        - name: tenantId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantUpdate"
      responses:
        "200":
          description: Updated tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"

  # ──────────────────────────────────────────────────────────
  # USERS
  # ──────────────────────────────────────────────────────────
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users in current tenant
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: User deleted

  # ──────────────────────────────────────────────────────────
  # ASSETS
  # ──────────────────────────────────────────────────────────
...
  /scans:
    get:
      operationId: listScans
      tags: [scans]
      summary: List all scans
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: assetId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of scans
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scan"
    post:
      operationId: createScan
      tags: [scans]
      summary: Create and schedule a scan
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScanInput"
      responses:
        "201":
          description: Scan created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{scanId}:
    get:
      operationId: getScan
      tags: [scans]
      summary: Get scan details
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
    delete:
      operationId: deleteScan
      tags: [scans]
      summary: Cancel or delete a scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /scans/{scanId}/cancel:
    post:
      operationId: cancelScan
      tags: [scans]
      summary: Cancel a running scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
  /scans/{scanId}/jobs:
    get:
      operationId: listScanJobs
      tags: [scans]
      summary: List scan jobs for a scan
      parameters:
        - name: scanId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scan jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScanJob"

  # ──────────────────────────────────────────────────────────
  # FINDINGS
  # ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, email, firstName, lastName, role, tenantId, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        tenantId:
          type: integer
        isActive:
          type: boolean
        lastLoginAt:
          type: ["string", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, firstName, lastName, role, password]
      properties:
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        password:
          type: string
    UserUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        role:
          type: string
        isActive:
          type: boolean

    # ── Tenant ────────────────────────────────────────────
    Tenant:
      type: object
      required: [id, name, slug, plan, isActive, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        isActive:
          type: boolean
        maxAssets:
          type: ["integer", "null"]
        maxUsers:
          type: ["integer", "null"]
        createdAt:
          type: string
    TenantInput:
      type: object
      required: [name, slug, plan]
      properties:
        name:
          type: string
        slug:
          type: string
        plan:
          type: string
        maxAssets:
          type: integer
        maxUsers:
          type: integer
    TenantUpdate:
      type: object
      properties:
        name:
          type: string
        plan:
          type: string
        isActive:
          type: boolean
        maxAssets:
          type: integer
        maxUsers:
          type: integer

    # ── Asset ─────────────────────────────────────────────
...
    Scan:
      type: object
      required: [id, tenantId, name, type, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        type:
          type: string
        status:
          type: string
        schedule:
          type: ["string", "null"]
        assetIds:
          type: array
          items:
            type: integer
        findingsCount:
          type: integer
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string
    ScanInput:
      type: object
      required: [name, type, assetIds]
      properties:
        name:
          type: string
        type:
          type: string
        assetIds:
          type: array
          items:
            type: integer
        schedule:
          type: string
        scanModules:
          type: array
          items:
            type: string
    ScanJob:
      type: object
      required: [id, scanId, assetId, status, createdAt]
      properties:
        id:
          type: integer
        scanId:
          type: integer
        assetId:
          type: integer
        status:
          type: string
        result:
          type: ["string", "null"]
        errorMessage:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        createdAt:
          type: string

    # ── Finding ───────────────────────────────────────────
...